get-sources: auto-convert .git repositories
A handful of packages are built from tags, branches or commits in Git
repositories. The way Pacman works is to bundle the entire .git
repository in the source packages for such components.

This is not only wasteful, it also makes it a bit harder to inspect
code.

Let's just re-package the respective revisions into individual .zip
packages. To make things super-convenient, also edit the PKGBUILD file
so that it uses the re-packaged .zip file.

Currently Git for Windows uses 10 components which are affected by this:
crypt, git-flow, mingw-w64-git, mingw-w64-libtre-git,
mingw-w64-rtmpdump-git, mingw-w64-winpthreads-git, mingw-w64-zlib,
msys2-runtime, rebase, and ssh-pageant-git.

All but one of these packages list the used Git repository as first
source entry, mingw-w64-zlib being the exception: it lists the minizip
Git repository as second entry (the first entry does not refer to a Git
repository).

Speaking of mingw-w64-zlib's first source entry: it is a URL. This is a
bit of a bummer because the zlib-<version>.tar.gz file referenced by
this URL is actually already included in the source package.

Therefore, while we are in the vicinity anyway, also edit those URL
entries in all PKGBUILD files to refer to the already-downloaded and
included files instead.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>