please.sh upgrade git-lfs: fix sed expression to update folder
When the folder is not the top-level directory, it contains slashes,
which have special meaning in sed expressions and need to be escaped.

This problem did not occur before because `upgrade git-lfs` was only
used to upgrade to version 2.2.0 so far, and that version's .zip files
contained the git-lfs.exe files at the top-level directory.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>