Remove getent calls from /etc/bash.bashrc
Remove calls to the "getent" utility, which is not currently included in
Git for Windows, from the system /etc/bash.bashrc file. Its presence is
not usually a problem, but when Cygwin is installed and also in the PATH
then Git Bash will encounter the following error during startup (see
issue #1226).

   1 [main] getent (14752) C:\cygwin\bin\getent.exe: *** fatal error - cygheap base mismatch detected - 0x1802FF408/0x180304408.

Signed-off-by: Adric Norris <landstander668@gmail.com>