sdk-installer: ensure that Git for Windows' keyring is imported
This is required so that the SDK can be installed fully, now that we
sign each and every package in Git for Windows' Pacman repository.

As the post-install script of msys2-keyring destroys every previous
attempt to initialize Pacman's keyring, we need to install
msys2-keyring (and gnupg) specifically before adding Git for Windows'
keys.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>