make-file-list: prepare for a BusyBox-based MinGit version
The idea of BusyBox is to have a single executable serving minimal
implementations of most Unix utilities as well as a small Unix shell.

This is perfect for what MinGit tries to accomplish: be as small as
possible, while still providing all non-interactive functionality
required by applications wishing to use Git's functionality.

This patch only prepares the script that generates the list of files to
include in MinGit (or the installer or portable Git or...).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>