please.sh upgrade: learn about BusyBox-w32
The idea of building a MINGW version of BusyBox is, of course, to allow
MinGit to shrink even further *and* make it more robust by avoiding the
MSYS2 runtime whenever possible.

Eventually, we will hopefully even be able to use BusyBox to execute
Git's Unix shell scripts in Git for Windows, but that will require quite
a bit more work, as BusyBox-w32 is unaware of MSYS2's tty emulation, and
would hence fail to consider Git Bash a valid interactive terminal.

Be that as it may, in the meantime we really need to automate the
process to upgrade the mingw-w64-busybox packages, as steam will most
likely pick up on that front.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>