pacman-helper: generate mingw32/mingw64 database packages, too
There is a very important distinction between MSYS and MINGW packages:
while it is possible to mix and match architectures in the latter (e.g.
have both 32-bit MINGW git.exe and 64-bit MINGW git.exe), the same is not
true of the former, due to MSYS binaries being tightly dependent on the
MSYS2 runtime (which is a single architecture, of course).

For that reason, the MSYS2 project has 32-bit/64-bit pairs for MSYS and
MINGW packages each, totaling 4 distinct Pacman repositories.

Git for Windows chose instead to have only a 32-bit and a 64-bit Pacman
repository, i.e. only two Pacman repositories.

That poses a problem, though, e.g. when building the mingw-w64-git
package: it depends on mingw-w64-curl, and Git for Windows overrides
MSYS2's mingw-w64-curl packages, and mingw-w64-git is built in both 32-bit
and 64-bit versions in one go within a 64-bit Git for Windows SDK (which
did not see the overridden 32-bit cURL package).

To remedy that, let's upload *two* versions of the package database to
each of the 32-bit and 64-bit Pacman repositories: one database having all
packages, as before, and one having only the MINGW packages.

That way, the 64-bit Git for Windows SDK can add the 64-bit repository and
the 32-bit *MINGW* repository, and everything is groovy again.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>