pacman-helper.sh push: handle failed database upload gracefully
It is possible that BinTray's API times out in the middle of a database
upload. When that occurs, it is a rather unfortunate problem, as we cannot
really update files in-place; We have to delete the database package
entirely and upload a new version instead. So when the upload failed,
there simply is no valid database.

In that case, the db_version function returns an empty string, and we
really want to try our best to fix the issue as fast as possible.

So let's shortcut everything and let `pacman-helper.sh push`

1) work in that case, and

2) skip determining which packages need to be uploaded

As a large chunk of code had to be indented, this commit is best viewed
with git diff's -w option.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>