git-extra: register "other" architecture's MINGW Pacman repository
Unlike MSYS2 packages, MINGW packages can be installed in both 32-bit and
64-bit versions at the same time, as they install into different
locations (and do not depend on the MSYS2 runtime).

With this change, Git for Windows' 32-bit and 64-bit SDKs learn to access
Git for Windows' MINGW packages for the "other" architecture, i.e. it is
now possible to install Git for Windows' mingw-w64-i686-curl and
mingw-w64-x86_64-curl packages at the same time.

This is now really necessary, too, as Git for Windows just started
shipping with a custom cURL version that supports choosing the SSL backend
at runtime, and to make use of this feature, mingw-w64-git (which is built
in both 32-bit and 64-bit versions in Git for Windows' 64-bit SDK)
requires the cURL version's headers offering the API function
curl_global_sslset().

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>