git-update: overhaul script a bit more
This imitates Git's scripts' argument handling, and handles the easy
up-to-date case first, saving a level of indentation for the remainder of
the function.

Still to do: determine PIDs referring to C:\Program Files\Git (or wherever
Git is installed), warn users about programs that are still running other
than the current script, and only kill them when the user gave their
consent.

Also to do: hook this up into the Start Menu, and maybe offer a profile.d/
script that tests up-to-date'ness upon startup of the Git Bash (but only
once per 24h at most).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>