installer/portable: exclude useless Python script even in 32-bit
The nghttpx tool tries to use a Python script for OCSP stapling support,
apparently. It is pulled into Git/portable Git/MinGit by virtue of
nghttp2 being a dependency of cURL, but cURL does not use that script.

Neither would it work in general, as it is a Python script and we do not
ship a Python interpreter.

That's why we excluded this script from being packaged in the
installer/portable Git/MinGit/NuGet/etc. By mistake, this inclusion only
extended to the 64-bit versions, now it includes the 32-bit versions,
too.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>