please.sh: remove obsolete workaround
When building Git, we really need the asciidoctor extensions, and for
technical reasons, we need them in 64-bit and 32-bit versions.

Earlier, our Pacman repositories had only one package database each, one
with all the 64-bit packages and one with all the 32-bit packages. And
since MSYS2 packages cannot be mixed and matched between architectures,
we required an ugly hack to install the 32-bit asciidoctor extensions
into a 64-bit SDK (because Git -- as all MINGW packages -- are built in
the 64-bit SDK for both architectures).

Now each repository has *two* package databases, one including all
packages for the given architecture, and one only including the MINGW
packages for said architecture. And Git for Windows' SDK accesses the
full database for the matching architecture, and the MINGW-only database
for the other architecture.

Therefore, the ugly hack is no longer required, as we can easily install
mingw-w64-i686-asciidoctor-extensions even in a 64-bit Git for Windows
SDK.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>