installer: fix auto-detection of symlink permissions
We added some code to auto-detect whether regular users can create
symbolic links (not only administrators).

Unfortunately, the detection was flawed, as we first tested whether the
installer is running as admin: we require it to be elevated.

So let's rethink this and use the hint from
https://stackoverflow.com/a/11995662 to detect whether the *original*
user (i.e. the user who started the installer) is an admin, before
testing symlink creation as that original user.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>