installer: override previous default of symlink option
Previous defaults were flawed by an incorrect test, and by the fact that
Windows' Developer Mode allows non-admin symlink creation only since
Creators Update. So it makes sense to change the default, which is made
more difficult by the fact that previous choices are remembered (with no
way of telling whether it was simply the default, or an explicit user
choice).

So let's override it if we detect that the previous choice was the
default option.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>