please.sh test_remote_branch: handle submodule gracefully
When the script detects that Git was checked out with core.autocrlf
set to `true`, it tries to check out the files again, with
core.autocrlf set to `false`.

To that end, it uses `git stash`. But that command has a bug currently
where it tries to ignore changes in submodules but then fails to do so.

Let's work around that by using `git reset --hard` instead.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>