git-extra: fix she-bang lines in /mingw*/bin/
When Ruby gems install shell scripts, on Windows they use absolute
Windows paths in the she-bang lines.

This is not what we want in the SDK, as it is supposed to be portable.

Happily, the she-bang lines are interpreted by Bash, which understands
"absolute Unix-style paths" that are in reality relative to the
top-level directory of the MSYS2 installation.

So let's convert the Windows paths in the she-bang lines.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>