installer: do not register `git update` scheduled task by default
It would most likely annoy users who do *not* want an auto-updater.

Besides, the current way to check for an update would flash text
windows when no update is available, i.e. most of the time. We might
want to revamp that script at some stage so that it runs without any
visible window, and uses, say, `git gui--askyesno` when the script
detected that an update is available.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>