installer: check again for blocking processes before complaining
Just before installing, we check for processes that still use files we
want to overwrite, and show the Processes page if such processes were
found.

When that happens, users typically close the corresponding windows and
then come back to the installer window.

While it would be nice to update the list when the focus is regained,
InnoSetup unfortunately lacks support for that right now.

So what happens is that the user clicks `Install` and gets a bogus
warning about the processes that need to be done before we can proceed,
forcing the user to click the `Refresh` button first.

What we *can* do, however, is to simply update the processes list right
before we would complain about the blocking processes, as if the user
had clicked the `Refresh` button first. So let's do that.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>