git-update: use a much more accurate method to count other Bash instances
When updating, a crucial bit of information for the user is: how many
Bash instances (other than the current one, if any) are running and
would need to be terminated to allow for Git for Windows to be updated?

This is an amazingly easy question to ask, yet an amazingly involved
question to answer.

For one, `git update` might have been called in Git CMD, in which case
there is no calling Bash at all. The previous logic would have warned
that it would be "killing -1 other Bash" if no Git Bash was running,
and not warn about killing a Bash instance when exactly one Git Bash
instance was active.

Further, it is a bit tricky to determine the current Bash instance, if
there is one, as MSYS2's idiosyncracies dictate that MSYS2 processes
spawned from pure Win32 processes report PPID 1 (and you have to follow
the PGID to find the spawning process instead).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>