please.sh upgrade: do not assume that @{upstream} is set
In automated builds, it may not be set at all. But then, we specifically
ensure that we can push to *origin*, and we ensure that the current
branch is `master` before pushing, so we may just as well go the whole
nine yards and hard-code that we want to push `master` to `origin`, but
only if it has local changes.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>