please.sh upgrade: improve warning when not on `master`
When running the `upgrade` operation on a branch other than `master`, we
do not push automatically. We may just as well say what branch we were
on, to give the user the information that she would need to know anyway.

Besides, this breaks up another long line to be shorter than 80 columns
per line.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>