please.sh upgrade msys2-runtime: error out early when target is in use
We cannot install the msys2-runtime package when msys-2.0.dll is in use.
For that reason, we error out when trying to install it.

When upgrading the msys2-runtime package, we will need to install it
after building it. It is utterly frustrating to wait until the
msys2-runtime is built before finding out that we cannot install it
because the dll is in use. So let's error out early.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>