please.sh upgrade: adjust debugging code even more
Even the previous test (which assumed that something funny is going on
with regards to the `mention` command) did not shed any light into the
build job failure to push the newest ReleaseNotes.

So let's revert the test item and rearrange the debugging code to allow
more diagnosing the next time a package needs to be upgraded.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>