please.sh upgrade: remove debugging code
The issue is resolved.

It is a bit embarrassing to note the root cause of the build job's
failure to push the latest release notes update after uploading a new
package, but oh well, here it goes.

The build job is running with a pre-populated .ssh/known_hosts, and an
earlier version of the build job failed to write a trailing LF to that
file. The first push, the one to update the PKGBUILD file, succeeded
because the file could be read, the line with the known host was parsed
alright.

Afterwards ssh wrote another line to that file because it just so
happened that the DNS lookup returned a different IP than was recorded
previously (.ssh/known_hosts' lines contain both host name and IP).

And here is where things went wrong: ssh did not realize that the file
ended in an unterminated line earlier, and blindly appended the new line
*without* writing a Line Feed first.

As a consequence, the .ssh/known_hosts file now contained an invalid
line (smooshing two lines together), and subsequent ssh calls would
simply ignore that entry.

This was not spotted earlier because the first time the build job
succeeded: it built git-extra 1.1.185.1178d5a, pushing the update to the
PKGBUILD successfully, and since git-extra updates do not get their
respective Release Notes entry, no subsequent push was necessary, hence
it did not become obvious that the .ssh/known_hosts file had been
corrupted.

The build job is configured to start from a fresh state, and therefore
the next time it was triggered (to upgrade msys2-runtime to 2.9.0-2),
its initial .ssh/known_hosts was a fresh copy. But now the Release Notes
got an update, and the second push failed. (It was performed manually.)

Later, Git LFS was upgraded to version 2.3.0, again trying to push a
Release Notes update (after pushing the PKGBUILD update already), and
that second push failed also. (It was also performed manually.)

Some serious debugging, including changes on build-extra's master
branch, was necessary to figure this one out.

And this, ladies and other gentle people, concludes today's debugging
adventure.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>