please.sh prerelease: really do not get confused by drive prefixes
When pushing the tag into the Git repository in Git for Windows SDK's
/usr/src/MINGW-packages/mingw-w64-git/src/git/ directory, let's just
turn the path into a pseudo-Unix directory. That way, Git does not even
see the drive prefix, and hence cannot think the drive letter is a
protocol specifier.

An earlier attempt to fix this (by prefixing with file:///) did not work
reliably.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>