git-extra: optionally let git-update ask via a message box
The idea of auto-updating Git for Windows is a really compelling one, as
it allows users to stay up-to-date with minimal hassle.

However, we will want to avoid flashing a CMD window in the common case
that the script determines that everything is up-to-date.

This window is actually necessary because the question whether to update
now (in the less common case that an update is available) is printed to
that window.

So let's add support for the --gui option, which lets the script ask the
user via git gui--askyesno instead. The next patch will update the
installer so that the scheduled task will avoid the CMD window
altogether.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>