installer: detect more in-use files
When any file in Git for Windows' installation is in use, an upgrade
will cause a dialog box "Access Denied ...".

Therefore, we already detect processes that use a couple of DLLs.

We detected in particular the MSYS2 processes. But we did not detect any
Git processes.

Ideally, we would detect *any* of Git for Windows' files that are in
use, but there seems to be no easy API to do that. So let's just try to
catch the most important entry points into Git for Windows (no user has
any business starting anything else in Git for Windows without going
through any of those entry points).

Pointed out by Matt Cavallari.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>