installer: do not activate autoupdate automatically when installed from VS
In discussions with a couple developers I trust, it became obvious that
these random popup dialog boxes may not be what unsuspecting users want
who merely installed Visual Studio with some payload that wants Git for
Windows.

Work is ongoing to come up with a design that is less intrusive, e.g.
mentioning after an interactive Git command, at most once a day, when an
update is available.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>