pacman-helper fetch: remove stale packages
The mingw-w64-asciidoctor-extensions package was updated recently,
and coincidentally, it was done in a fresh VM. A subsequent update was run
from the previously single VM where packages were built, and as a
consequence, a stale package was still there.

The problem occurred when that subsequent update simply tried to add all
*.pkg.tar.xz files, because *.pkg.tar.xz is not ordered by version order,
but alphabetically, and 150 is *smaller* than 79 in that latter order,
essentially forcing a downgrade.

Let's just make sure that a `fetch` operation not only fetches newer
packages, but also removes stale packages, so that the order simply won't
matter.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>