Add a script to estimate installation sizes
It is a valuable piece of information to know how much space a
particular Git for Windows version will occupy *on disk* once installed,
and a particularly difficult piece of information, too.

We can use the mingw-w64-innoextract package to list contents of
installers, which is useful, and we can hard-code an allocation unit
size of 4096 (which can be expected in most setups), but that still does
not fully account for the final size on disk, as it leaves out the
uninstaller, the size required for the directories and various files
that are generated/hardlinked upon installation.

To account for this difference, let's just hard-code the difference as
measured from a 64-bit Git for Windows v2.14.2 installation, and let's
just keep in mind that the resulting number is more of a best guess than
a precise number of bytes.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>