please.sh install git: re-install latest git-extra
Previously, everything went smoothly when we called pkg_install on
git-extra after installing mingw-w64-git, because everything was built
in the same VM.

This changed: now the idea is that the packages are built in Docker
containers.

So let's make sure that the *latest* git-extra is installed, not the one
we built on *this* machine last time.

This seems to be the culprit of the substantially larger portable Gits
in v2.14.2: https://github.com/git-for-windows/git/issues/1308

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>