pacman-helper push_missing_signatures: sync mingw to mingw&msys databases
The authoritative source for the current packages is the full, i.e.
MINGW & MSYS database, with the MINGW-only database (e.g.
git-for-windows-mingw32) being a strict subset.

Recently, when the mingw-w64-curl-winssl-bin packages were made
obsolete, however, I forgot to update the MINGW database.

Let's teach the push_missing_signatures command the trick to adjust the
MINGW-only database in case it contains left-over entries.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>