pacman-helper: work around upload timeouts
Unfortunately, BinTray times out during upload for this developer quite
frequently. Even worse, it seems to be correlated to the number of small
files, making it most likely to happen while uploading the package
databases, which incidentally is the worst time to time out because the
database needs to be available from a fixed file location, hence the
previous database version has to be deleted (and therefore, during that
window between deleting the previous database version and finishing the
upload of the new database, Git for Windows' Pacman repository is in an
invalid state and nobody can update their installations).

Another unfortunate fact is that cURL's --retry option seems not to
catch here, most likely because the timeout happens during the initial
handshake.

This means that we have to use all the dirty tricks with which we can
come up to avoid those timeouts. And the dirtiest trick of 'em all is to
just try again, and again, and again.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>