git-extra: add a utility to work with/around blocked files
A file is blocked when it is in use by some process, and cannot be
deleted in such a case.

This is the case, for example, when a user still has a Git Bash open
while trying to install a new Git for Windows version: the .exe and .dll
files that are in use by the Git Bash process and its children cannot be
overwritten.

This patch introduces a new utility that can be used to detect such a
problem, and as a workaround those files can also be moved into the
recycle bin (which is possible while they are in use).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>