please.sh publish: require w3m
Previously, the `release` and the `publish` subcommands have been
performed in the same dedicated VM.

The reason why those steps are separate is to allow for last-minute
manual "integration" testing of the generated installers.

This is very important, as a few bugs have been caught then and there.
For example, the first `release` run of Git for Windows v2.14.3 produced
an installer which exposed the dreaded post-install.bat bug (which was
subsequently fixed by merging PR git-for-windows/build-extra#160). The
bug, therefore, could be fixed before any user (other than myself) was
affected.

However, with Git for Windows v2.14.3, the release process was moved to
VSTS jobs, and the `release` and `publish` steps are now two separate
jobs, meaning that they can run in different build agents.

So let's make extra sure that all prerequisites are installed before
publishing the release.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>