installer: run auto-update maybe a little later than scheduled
When we install the Scheduled Task to trigger `git update` to run, we
did not check that checkbox "Run task as soon as possible after a
scheduled start is missed". But we should.

This is not possible via a command-line option of schtasks.exe,
therefore we have to switch to using an XML file (which would have been
a problem when we still supported Windows XP, but Vista and above accept
that XML with only a few required entries).

Pointed out by Matthew Cheetham.

While at it, we are now also able to ask the Scheduled Task only to be
run if there is a network connection (it is pointless to run it without
internet, right?).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>