installer: test for more in-use files
The recently-added "blocked-file-util" is now being put to good use: we
enumerate the process IDs of the processes that use .exe or .dll files
from the installation location to skip the installation if the
/SKIPIFINUSE=1 option was specified.

This catches scenarios where users call non-Git executables directly
(e.g. C:\Program Files\Git\mingw64\bin\openssl.exe) and a silent
installation wants to just make sure that Git is installed.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>