installer/portable: exclude dysfunctional /usr/bin/xml*.exe
We ship e.g. xmllint.exe, but that executable fails to start because we
do not ship the dependency msys-icuuc59.dll.

This is not a problem, as xmllint is absolutely unnecessary for Git for
Windows' functionality.

Therefore, let's just remove it. (Users who want a working xmllint.exe
can always install Git for Windows' SDK, as they are really more
interested in a system with package management than actual Git.)

This fixes https://github.com/git-for-windows/git/issues/1336

While at it, remove xmlcatalog.exe for the same reason. And xmlwf.exe
for good measure, because we really do not need it in Git.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>