please.sh: do not complain if 32-bit Git for Windows SDK does not exist
Or better put: do not complain if the location specified by
`windows.sdk32.path` does not exist (yet).

The reason: this script is in heavy use in VSTS build agents, where it
is thoroughly possible that only the 64-bit Git for Windows SDK is
initialized (e.g. when there is no need for the 32-bit one).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>