Add nano text editor
Nano is a small text editor that is -- for novices of the bash command
line -- easier to use than e.g. vim, as it displays shortcuts for common
tasks (write out, exit).

Adding it to the Git for Windows installer only increases the installer
size by ~100kB, which is acceptable for the benefit.

[jes: auto-install nano as needed when packaging a new installer]

Signed-off-by: Oliver Stueker <ostueker@mun.ca>
Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>