installer: turn `nano` option into a full wizard page
This allows us to

- describe the option better,

- mark it as recommended,

- describe the `vim` option better (including the fact that it is the
  default, and why), and

- provide hyperlinks where users can learn more about `nano` vs `vim`.

It also keeps the door wide open to add additional editor choices, e.g.
when the user has Notepad++ installed; it is likely their editor of
choice and much preferred over vim if it is installed. And we want to
keep this door open, see e.g. https://www.xkcd.com/378/

The way to force GNU nano to be used when running the installer via the
command-line is to use /o:EditorOption=Nano

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>