installer: refactor wizard page code
We have a lot of semi-duplicated code to add tons of radio buttons, and
most of the code is tedious to maintain and even more tedious is it to
add new code, as basically all of the dimensions have to be measured,
adjusted and tested.

Even worse, when rewording the description of an option, it frequently
changes the number of lines, which requires adjustments not only in that
option, but every subsequent option, too.

This is a job not for humans, but for the machine.

So let's introduce code to make it so. One function to create a new
wizard page. One function to add radio buttons and descriptions.

While at it, automate also the highly repetitive code to highlight parts
of the descriptions in red, and to turn other parts into hyperlinks.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>