please.sh upgrade: optionally force a pkgrel
There was a botched release of the OpenSSL 1.0.2m package: some
overzealous linter in /usr/share/makepkg/lint_package/ wanted to run
xargs, and for some stupid reason this failed with

	assertion "bc_ctl.arg_max >= LINE_MAX" failed:
	file "xargs.c", line 500, function: main

(but of course that did not abort the packaging, oh no)

Even more unfortunate was the fact that this failure produced
xargs.exe.stackdump files that were packaged.

To fix this, we now need to re-package this, with a bumped-up pkgrel so
that it works (and does not complain about the package already being
made).

We can do this now using the --force-pkgrel=2 option.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>