Mention OpenSSL 1.0.2m only once in the Release Notes
Sadly, the first attempt at packaging included the file
xargs.exe.stracktrace in the top-level directory of every package. Which
is not only annoying, it also prevents the packages from being installed
due to conflicts.

The second attempt worked, at last. But it also added another entry to
the release notes, because I used the automated `./please.sh upgrade`
command to perform the packaging.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>