git-extra: help avoid bogus xargs.exe.stackdump files in packages
When makepkg runs its linters, it uses a `find | xargs` invocation to
warn about files that include the source directory.

Sadly, it turns out that this invocation fails at least *sometimes* in
Docker containers used by yours truly to automate Git for Windows
maintenance tasks, with:

            assertion "bc_ctl.arg_max >= LINE_MAX" failed:
            file "xargs.c", line 500, function: main

and leaving xargs.exe.stackdump files. In the directory that gets
bundled up.

While it would be nicer to avoid that stackdump altogether, and even
better: to have xargs perform its job as planned, it would seem that
this is beyond our reach for now (why is it hitting that limit
*sometimes* but not every time?).

So let's do the next best thing and simply run another "linter" that
removes those bogus stackdumps.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>