git-extra: fix patch for vim's defaults.vim
A change in the defaults.vim file was not caught in time, and the patch
in git-extra's post-install script (which tries to prevent vim from
*forcing* to jump back to the line where we left off upon opening, even
commit messages, where it Don't Make No Sense) caused a problem by
commenting out waaaaaay too many lines.

Unfortunately, this change was so close to Git for Windows v2.15.1 that
it affects end users now.

Let's undo the damage in one big hurry.

This fixes https://github.com/git-for-windows/git/issues/1382

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>