please.sh: make a couple of pushes more resilient
In particular when using VSTS agents, it is quite possible that some
tasks work on Git repositories in parallel and want to commit and push.

That can lead to problems when one of the pushes does not fast-forward.

In some cases, the resolution is trivial. So much so that we can
automate it: simply pull and try to push again.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>