please.sh prerelease: bundle the git-wrapper in place of builtins
As Windows Explorer (and even Inno Setup itself) cannot determine the
size of Git for Windows correctly when there are hard-linked copies of
git.exe all over the place (overestimating a 243MB install to the tune
of 500MB, i.e. not even by a small amount), Git for Windows ships with
small .exe files in place of the hardlinked copies, the "git-wrapper".

The purpose of the "git-wrapper" was originally only to set up a couple
of environment variables before handing off to git.exe (preserving the
command-line parameters), and in this instance simply does the same
(slightly differently, though, as it has to extract the builtin's name
from the executable name and shift that name in argv[1]).

The trick to swap out the hard-linked builtins and swap in hard-linked
copies of the git-wrapper is performed by the post-install script of
the git-extra package.

Therefore, let's just install the git-extra package after installing the
mingw-w64-git package.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>