please.sh prerelease: allow resetting the "mode"
Currently, there are three supported modes: installer, portable and
mingit. But once they are enabled, there was no way to disable them
again on the same command-line.

Let's introduce --only-installer and friends, as well as --reset-mode,
to fix that.

This comes in particularly handy when reusing --now for publishing
MinGit-only pre-releases.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>