please.sh build/upgrade: optionally blow away the src/ directory
There has been a problem with the packaging of the OpenSSL 1.0.2n
package: its libeay32.dll was miraculously not updated, as witnessed by
its file version.

The reason seems to be that the source files of the library have not
changed since OpenSSL 1.0.2m, and therefore it had not been relinked
(the 1.0.2m/1.0.2n builds happened on the same machine).

This may very well be a bug in OpenSSL's build definition, but that does
not necessarily help us here: we cannot chase down every dependency's
build definition bugs.

So let's just use makepkg's --cleanbuild option to *enforce* a fresh
build. Make that an opt-in, though, to allow for the same development
speed as before, and use that flag only in the appropriate context.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>