please.sh upgrade --force git: ignore existing tags
In a CI build, it is quite possible that a previously-failed attempt at
building a Git package left a tag behind that would conflict with the
one we want to create. We already introduced the --force flag to force
re-tagging, but we failed to ignore the existing tag while updating the
release notes.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>