please.sh finalize: instead of ignoring a colliding tag, delete it
When a build in Visual Studio Team Services failed (e.g. because of the
stupid issue that was introduced by turning `sha1collisiondetection`
into a submodule... why???), we will want to re-run the build,
re-creating the same tag that we already created in the previous
(failed) run.

A previous attempt at fixing this tried to ignore the existing tag, but
that led to Git for Windows v2.16.0(2) being released (without any
v2.16.0...).

Let's just convert that option to delete any existing tag instead, so
that it can be safely re-created. And then still use that option in the
`upgrade` subcommand.

This fixes https://github.com/git-for-windows/git/issues/1433

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>