please.sh upgrade git: work around pesky Submodule problems
Ah, submodules. Don't get me started on submodules. The submodule
"feature" of Git is based on the erroneous assumption that repositories
checked out into a subdirectory of an existing worktree could be handled
in the exact same way as files.

Which is, uhm, how to say it, incorrect.

For example, when you have a submodule with untracked files, you cannot
simply delete it, then recreate it, and hope to have all files back.

Also, converting directories to submodules and back has problems. So
yes, we see exactly this problem in the Git for Windows build:

	error: sha1collisiondetection: is a directory - add files inside instead

Let's just go ahead and delete the sha1collisiondetection directory (if
it exists from previous runs) and be done with it. We did not cause the
problem, we should not be held accountable to clean things up.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>