versions: prefer InnoUnp over InnoExtract
While InnoExtract can be built from source in MSYS2 (because it is
written in C++, as opposed to InnoUnp, which is written using Delphi),
it has become really tedious to try to work around the Segmentation
Fault that is haunting literally every call (even with `-h`).

So let's just favor InnoUnp (http://innounp.sourceforge.net/).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>